+++ cardano-cli/cardano-cli.cabal
@@ -106,7 +106,6 @@ library
                         Cardano.CLI.Legacy.Run.Node
                         Cardano.CLI.Legacy.Run.Pool
                         Cardano.CLI.Legacy.Run.Query
-                        Cardano.CLI.Legacy.Run.Read
                         Cardano.CLI.Legacy.Run.StakeAddress
                         Cardano.CLI.Legacy.Run.TextView
                         Cardano.CLI.Legacy.Run.Transaction
@@ -116,6 +115,7 @@ library
                         Cardano.CLI.OS.Posix
                         Cardano.CLI.Parser
                         Cardano.CLI.Pretty
+                        Cardano.CLI.Read
                         Cardano.CLI.Render
                         Cardano.CLI.Run
                         Cardano.CLI.Run.Ping
