+++ cardano-cli/src/Cardano/CLI/Legacy/Run/Governance.hs
@@ -15,7 +15,7 @@ import qualified Cardano.Api.Shelley as Api
 import           Cardano.CLI.Commands.Governance
 import           Cardano.CLI.EraBased.Run.Governance
 import           Cardano.CLI.Legacy.Commands.Governance
-import           Cardano.CLI.Legacy.Run.Read (fileOrPipe, readFileTx)
+import           Cardano.CLI.Read (fileOrPipe, readFileTx)
 import           Cardano.CLI.Types.Common
 import           Cardano.CLI.Types.Governance
 import qualified Cardano.CLI.Types.Governance as Cli
