+++ cardano-cli/src/Cardano/CLI/Legacy/Run/Transaction.hs
@@ -24,8 +24,8 @@ import           Cardano.Api.Shelley
 import           Cardano.CLI.Json.Friendly (friendlyTxBS, friendlyTxBodyBS)
 import           Cardano.CLI.Legacy.Commands.Transaction
 import           Cardano.CLI.Legacy.Run.Genesis
-import           Cardano.CLI.Legacy.Run.Read
 import           Cardano.CLI.Legacy.Run.Validate
+import           Cardano.CLI.Read
 import           Cardano.CLI.Types.Common
 import           Cardano.CLI.Types.Governance
 import           Cardano.CLI.Types.Output
